+++ applications/fishing-map/src/features/map/Map.tsx
@@ -1,4 +1,5 @@
 import React, { useCallback, useEffect, useState, useMemo } from 'react'
+import { AsyncReducerStatus } from 'types'
 import { createPortal } from 'react-dom'
 import { MiniGlobe, IconButton, MiniglobeBounds } from '@globalfishingwatch/ui-components'
 import { InteractiveMap, ScaleControl, MapRequest } from '@globalfishingwatch/react-map-gl'
@@ -67,7 +68,7 @@ const Map = (): React.ReactElement => {
     [token]
   )
 
-  const { clickedEvent, dispatchClickedEvent } = useClickedEventConnect()
+  const { clickedEvent, clickedEventStatus, dispatchClickedEvent } = useClickedEventConnect()
   const onMapClick = useMapClick(dispatchClickedEvent)
   const clickedTooltipEvent = useMapTooltip(clickedEvent)
   const closePopup = useCallback(() => {
@@ -116,7 +117,13 @@ const Map = (): React.ReactElement => {
           <div className={styles.scale}>
             <ScaleControl maxWidth={100} unit="nautical" />
           </div>
-          {clickedEvent && <ClickPopup event={clickedTooltipEvent} onClose={closePopup} />}
+          {clickedEvent && (
+            <ClickPopup
+              event={clickedTooltipEvent}
+              onClose={closePopup}
+              loading={clickedEventStatus === AsyncReducerStatus.Loading}
+            />
+          )}
         </InteractiveMap>
       )}
       <div className={styles.mapControls}>