+++ applications/fishing-map/src/features/workspace/workspace.hook.ts
@@ -3,11 +3,14 @@ import { useCallback } from 'react'
 import { WorkspaceDataviewConfig } from '@globalfishingwatch/dataviews-client'
 import { selectDataviewsConfig } from 'routes/routes.selectors'
 import { useLocationConnect } from 'routes/routes.hook'
+import { selectWorkspaceDataviewConfig } from './workspace.selectors'
 
 export const useDataviewsConfigConnect = () => {
   const urlDataviewsConfig = useSelector(selectDataviewsConfig)
+  const workspaceDataviewsConfig = useSelector(selectWorkspaceDataviewConfig)
   const { dispatchQueryParams } = useLocationConnect()
-  const updateUrlDataviewConfig = useCallback(
+
+  const updateDataviewConfig = useCallback(
     (dataviewConfig: Partial<WorkspaceDataviewConfig>) => {
       const currentDataviewConfig = urlDataviewsConfig?.find(
         (urlDataviewConfig) => urlDataviewConfig.id === dataviewConfig.id
@@ -27,5 +30,21 @@ export const useDataviewsConfigConnect = () => {
     },
     [dispatchQueryParams, urlDataviewsConfig]
   )
-  return { updateUrlDataviewConfig }
+
+  const deleteDataviewConfig = useCallback(
+    (id: string) => {
+      const dataviewsConfig = (urlDataviewsConfig || []).filter(
+        (urlDataviewConfig) => urlDataviewConfig.id !== id
+      )
+      const workspaceDataviewConfig = workspaceDataviewsConfig?.find(
+        (dataviewConfig) => dataviewConfig.id === id
+      )
+      if (workspaceDataviewConfig) {
+        dataviewsConfig.push({ id, deleted: true })
+      }
+      dispatchQueryParams({ dataviewsConfig })
+    },
+    [dispatchQueryParams, urlDataviewsConfig, workspaceDataviewsConfig]
+  )
+  return { updateDataviewConfig, deleteDataviewConfig }
 }