+++ applications/amathea/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @globalfishingwatchapp/amathea
 
+## 0.11.1
+
+### Patch Changes
+
+- fae4346: Update types
+- Updated dependencies [fae4346]
+  - @globalfishingwatch/dataviews-client@3.1.1
+  - @globalfishingwatch/layer-composer@2.13.4
+
 ## 0.11.0
 
 ### Minor Changes